+++ src/components/App.tsx
@@ -2,7 +2,7 @@ import React, { useContext } from "react";
 import { RoughCanvas } from "roughjs/bin/canvas";
 import rough from "roughjs/bin/rough";
 import clsx from "clsx";
-import { supported as fsSupported } from "browser-fs-access";
+import { fileOpen, supported as fsSupported } from "browser-fs-access";
 import { nanoid } from "nanoid";
 
 import {
@@ -53,6 +53,7 @@ import {
   ENV,
   EVENT,
   GRID_SIZE,
+  IMAGE_RENDER_TIMEOUT,
   LINE_CONFIRM_THRESHOLD,
   MIME_TYPES,
   MQ_MAX_HEIGHT_LANDSCAPE,
@@ -92,6 +93,7 @@ import {
   newElement,
   newLinearElement,
   newTextElement,
+  newImageElement,
   textWysiwyg,
   transformElements,
   updateTextElement,
@@ -115,6 +117,7 @@ import { deepCopyElement, newFreeDrawElement } from "../element/newElement";
 import {
   isBindingElement,
   isBindingElementType,
+  isInitializedImageElement,
   isLinearElement,
   isLinearElementType,
 } from "../element/typeChecks";
@@ -126,6 +129,8 @@ import {
   ExcalidrawLinearElement,
   ExcalidrawTextElement,
   NonDeleted,
+  InitializedExcalidrawImageElement,
+  ExcalidrawImageElement,
 } from "../element/types";
 import { getCenter, getDistance } from "../gesture";
 import {
@@ -141,9 +146,9 @@ import History from "../history";
 import { defaultLang, getLanguage, languages, setLanguage, t } from "../i18n";
 import {
   CODES,
-  getResizeCenterPointKey,
-  getResizeWithSidesSameLengthKey,
-  getRotateWithDiscreteAngleKey,
+  shouldResizeFromCenter,
+  shouldMaintainAspectRatio,
+  shouldRotateWithDiscreteAngle,
   isArrowKey,
   KEYS,
 } from "../keys";
@@ -166,6 +171,7 @@ import { SceneState, ScrollBars } from "../scene/types";
 import { getNewZoom } from "../scene/zoom";
 import { findShapeByKey } from "../shapes";
 import {
+  AppClassProperties,
   AppProps,
   AppState,
   ExcalidrawImperativeAPI,
@@ -196,6 +202,13 @@ import LayerUI from "./LayerUI";
 import { Stats } from "./Stats";
 import { Toast } from "./Toast";
 import { actionToggleViewMode } from "../actions/actionToggleViewMode";
+import { generateIdFromFile, isImageFile } from "../data/blob";
+import {
+  getInitializedImageElements,
+  updateImageCache,
+} from "../element/image";
+import throttle from "lodash.throttle";
+import { getDataURL } from "../excalidraw-app/data";
 
 const IsMobileContext = React.createContext(false);
 export const useIsMobile = () => useContext(IsMobileContext);
@@ -226,7 +239,7 @@ const gesture: Gesture = {
 };
 
 class App extends React.Component<AppProps, AppState> {
-  canvas: HTMLCanvasElement | null = null;
+  canvas: AppClassProperties["canvas"] = null;
   rc: RoughCanvas | null = null;
   unmounted: boolean = false;
   actionManager: ActionManager;
@@ -243,7 +256,7 @@ class App extends React.Component<AppProps, AppState> {
   private scene: Scene;
   private resizeObserver: ResizeObserver | undefined;
   private nearestScrollableContainer: HTMLElement | Document | undefined;
-  public library: Library;
+  public library: AppClassProperties["library"];
   public libraryItemsFromStorage: LibraryItems | undefined;
   private id: string;
   private history: History;
@@ -251,6 +264,8 @@ class App extends React.Component<AppProps, AppState> {
     container: HTMLDivElement | null;
     id: string;
   };
+
+  public imageCache: AppClassProperties["imageCache"] = new Map();
 
   constructor(props: AppProps) {
     super(props);
@@ -287,6 +302,7 @@ class App extends React.Component<AppProps, AppState> {
         ready: true,
         readyPromise,
         updateScene: this.updateScene,
+        setFiles: this.setFiles,
         resetScene: this.resetScene,
         getSceneElementsIncludingDeleted: this.getSceneElementsIncludingDeleted,
         history: {
@@ -445,6 +461,7 @@ class App extends React.Component<AppProps, AppState> {
               focusContainer={this.focusContainer}
               library={this.library}
               id={this.id}
+              onImageAction={this.onImageAction}
             />
             <div className="excalidraw-textEditorContainer" />
             <div className="excalidraw-contextMenuContainer" />
@@ -470,7 +487,7 @@ class App extends React.Component<AppProps, AppState> {
     );
   }
 
-  public focusContainer = () => {
+  public focusContainer: AppClassProperties["focusContainer"] = () => {
     if (this.props.autoFocus) {
       this.excalidrawContainerRef.current?.focus();
     }
@@ -702,6 +719,10 @@ class App extends React.Component<AppProps, AppState> {
     const scene = restore(initialData, null, null);
     scene.appState = {
       ...scene.appState,
+      elementType:
+        scene.appState.elementType === "image"
+          ? "selection"
+          : scene.appState.elementType,
       isLoading: false,
     };
     if (initialData?.scrollToContent) {
@@ -726,6 +747,11 @@ class App extends React.Component<AppProps, AppState> {
       ...scene,
       commitToHistory: true,
     });
+
+    this.refreshImages(
+      getInitializedImageElements(scene.elements),
+      scene.appState.files,
+    );
 
     const libraryUrl =
       // current
@@ -817,6 +843,7 @@ class App extends React.Component<AppProps, AppState> {
   }
 
   public componentWillUnmount() {
+    this.imageCache.clear();
     this.resizeObserver?.disconnect();
     this.unmounted = true;
     this.removeEventListeners();
@@ -1027,17 +1054,26 @@ class App extends React.Component<AppProps, AppState> {
       );
       cursorButton[socketId] = user.button;
     });
-    const elements = this.scene.getElements();
+    const renderingElements = this.scene.getElements().filter((element) => {
+      if (element.type === "image") {
+        if (
+          // not placed on canvas yet (but in elements array)
+          this.state.pendingImageElement &&
+          element.id === this.state.pendingImageElement.id
+        ) {
+          return false;
+        }
+      }
+      // don't render text element that's being currently edited (it's
+      // rendered on remote only)
+      return (
+        !this.state.editingElement ||
+        this.state.editingElement.type !== "text" ||
+        element.id !== this.state.editingElement.id
+      );
+    });
     const { atLeastOneVisibleElement, scrollBars } = renderScene(
-      elements.filter((element) => {
-        // don't render text element that's being currently edited (it's
-        // rendered on remote only)
-        return (
-          !this.state.editingElement ||
-          this.state.editingElement.type !== "text" ||
-          element.id !== this.state.editingElement.id
-        );
-      }),
+      renderingElements,
       this.state,
       this.state.selectionElement,
       window.devicePixelRatio,
@@ -1054,6 +1090,8 @@ class App extends React.Component<AppProps, AppState> {
         remotePointerUsernames: pointerUsernames,
         remotePointerUserStates: pointerUserStates,
         shouldCacheIgnoreZoom: this.state.shouldCacheIgnoreZoom,
+        theme: this.state.theme,
+        imageCache: this.imageCache,
       },
       {
         renderOptimizations: true,
@@ -1067,12 +1105,14 @@ class App extends React.Component<AppProps, AppState> {
       // hide when editing text
       this.state.editingElement?.type === "text"
         ? false
-        : !atLeastOneVisibleElement && elements.length > 0;
+        : !atLeastOneVisibleElement && renderingElements.length > 0;
     if (this.state.scrolledOutside !== scrolledOutside) {
       this.setState({ scrolledOutside });
     }
 
     this.history.record(this.state, this.scene.getElementsIncludingDeleted());
+
+    this.scheduleImageRefresh();
 
     // Do not notify consumers if we're still loading the scene. Among other
     // potential issues, this fixes a case where the tab isn't focused during
@@ -1193,6 +1233,22 @@ class App extends React.Component<AppProps, AppState> {
       ) {
         return;
       }
+
+      const file = event?.clipboardData?.files[0];
+      if (isImageFile(file)) {
+        const { x: sceneX, y: sceneY } = viewportCoordsToSceneCoords(
+          { clientX: cursorX, clientY: cursorY },
+          this.state,
+        );
+
+        const imageElement = this.createImageElement({ sceneX, sceneY });
+        this.insertImageElement(imageElement, file);
+        this.initializeImageDimensions(imageElement);
+        this.setState({ selectedElementIds: { [imageElement.id]: true } });
+
+        return;
+      }
+
       const data = await parseClipboard(event);
       if (this.props.onPaste) {
         try {
@@ -1389,8 +1445,12 @@ class App extends React.Component<AppProps, AppState> {
     maxZoom: number = 1, //null will zoom to max based on viewport
     margin: number = 0.03, //percentage of viewport width&height
   ) => {
-    if(!target) target = this.scene.getElements();
-    if(target.length===0) maxZoom = 1;
+    if (!target) {
+      target = this.scene.getElements();
+    }
+    if (target.length === 0) {
+      maxZoom = 1;
+    }
     this.setState(
       zoomToFitElements(target, this.state, false, maxZoom, margin).appState,
     );
@@ -1419,6 +1479,25 @@ class App extends React.Component<AppProps, AppState> {
       this.setState({ errorMessage: error.message });
     }
   };
+
+  public setFiles: ExcalidrawImperativeAPI["setFiles"] = withBatchedUpdates(
+    (files) => {
+      this.setState(
+        (state) => ({
+          files: {
+            ...state.files,
+            ...files.reduce((acc, { id, type, dataURL }) => {
+              acc[id] = { type, id, dataURL };
+              return acc;
+            }, {} as Record<string, AppState["files"][number]>),
+          },
+        }),
+        () => {
+          this.refreshImages();
+        },
+      );
+    },
+  );
 
   public updateScene = withBatchedUpdates(
     <K extends keyof AppState>(sceneData: {
@@ -1510,7 +1589,7 @@ class App extends React.Component<AppProps, AppState> {
         this.setState({ isBindingEnabled: false });
       }
 
-      if (event.code === CODES.NINE) {
+      if (event.code === CODES.ZERO) {
         this.setState({ isLibraryOpen: !this.state.isLibraryOpen });
       }
 
@@ -1673,6 +1752,9 @@ class App extends React.Component<AppProps, AppState> {
     }
     if (!isLinearElementType(elementType)) {
       this.setState({ suggestedBindings: [] });
+    }
+    if (elementType === "image") {
+      this.onImageAction();
     }
     if (elementType !== "selection") {
       this.setState({
@@ -2389,6 +2471,26 @@ class App extends React.Component<AppProps, AppState> {
         this.state.elementType,
         pointerDownState,
       );
+    } else if (this.state.elementType === "image") {
+      // reset image preview on pointerdown
+      setCursor(this.canvas, CURSOR_TYPE.CROSSHAIR);
+
+      if (!this.state.pendingImageElement) {
+        return;
+      }
+
+      this.setState({
+        draggingElement: this.state.pendingImageElement,
+        editingElement: this.state.pendingImageElement,
+        pendingImageElement: null,
+        multiElement: null,
+      });
+
+      const { x, y } = viewportCoordsToSceneCoords(event, this.state);
+      mutateElement(this.state.pendingImageElement, {
+        x,
+        y,
+      });
     } else if (this.state.elementType === "freedraw") {
       this.handleFreeDrawElementOnPointerDown(
         event,
@@ -2958,6 +3060,32 @@ class App extends React.Component<AppProps, AppState> {
     });
   };
 
+  private createImageElement = ({
+    sceneX,
+    sceneY,
+  }: {
+    sceneX: number;
+    sceneY: number;
+  }) => {
+    const [gridX, gridY] = getGridPoint(sceneX, sceneY, this.state.gridSize);
+
+    const element = newImageElement({
+      type: "image",
+      x: gridX,
+      y: gridY,
+      strokeColor: this.state.currentItemStrokeColor,
+      backgroundColor: this.state.currentItemBackgroundColor,
+      fillStyle: this.state.currentItemFillStyle,
+      strokeWidth: this.state.currentItemStrokeWidth,
+      strokeStyle: this.state.currentItemStrokeStyle,
+      roughness: this.state.currentItemRoughness,
+      opacity: this.state.currentItemOpacity,
+      strokeSharpness: this.state.currentItemLinearStrokeSharpness,
+    });
+
+    return element;
+  };
+
   private handleLinearElementOnPointerDown = (
     event: React.PointerEvent<HTMLCanvasElement>,
     elementType: ExcalidrawLinearElement["type"],
@@ -3342,7 +3470,7 @@ class App extends React.Component<AppProps, AppState> {
         let dx = gridX - draggingElement.x;
         let dy = gridY - draggingElement.y;
 
-        if (getRotateWithDiscreteAngleKey(event) && points.length === 2) {
+        if (shouldRotateWithDiscreteAngle(event) && points.length === 2) {
           ({ width: dx, height: dy } = getPerfectElementSize(
             this.state.elementType,
             dx,
@@ -3519,6 +3647,10 @@ class App extends React.Component<AppProps, AppState> {
         pointerDownState.eventListeners.onKeyUp!,
       );
 
+      if (this.state.pendingImageElement) {
+        this.setState({ pendingImageElement: null });
+      }
+
       if (draggingElement?.type === "freedraw") {
         const pointerCoords = viewportCoordsToSceneCoords(
           childEvent,
@@ -3546,6 +3678,27 @@ class App extends React.Component<AppProps, AppState> {
 
         this.actionManager.executeAction(actionFinalize);
 
+        return;
+      }
+      if (draggingElement?.type === "image") {
+        const imageElement = draggingElement;
+        try {
+          this.initializeImageDimensions(imageElement);
+          this.setState(
+            { selectedElementIds: { [imageElement.id]: true } },
+            () => {
+              this.actionManager.executeAction(actionFinalize);
+            },
+          );
+        } catch (error) {
+          console.error(error);
+          this.scene.replaceAllElements(
+            this.scene
+              .getElementsIncludingDeleted()
+              .filter((el) => el.id !== imageElement.id),
+          );
+          this.actionManager.executeAction(actionFinalize);
+        }
         return;
       }
 
@@ -3782,6 +3935,249 @@ class App extends React.Component<AppProps, AppState> {
     });
   }
 
+  private initializeImage = async ({
+    imageFile,
+    imageElement: _imageElement,
+  }: {
+    imageFile: File;
+    imageElement: ExcalidrawImageElement;
+  }) => {
+    // generate image id (digest) before any resizing/compression takes place to
+    // keep it more portable
+    const imageId = await generateIdFromFile(imageFile);
+
+    const dataURL =
+      this.state.files[imageId]?.dataURL || (await getDataURL(imageFile));
+
+    const imageElement = mutateElement(
+      _imageElement,
+      {
+        imageId,
+      },
+      false,
+    ) as NonDeleted<InitializedExcalidrawImageElement>;
+
+    return new Promise<NonDeleted<InitializedExcalidrawImageElement>>(
+      (resolve, reject) => {
+        this.setState(
+          (state) => ({
+            files: {
+              ...state.files,
+              [imageId]: {
+                type: "image",
+                id: imageId,
+                dataURL,
+              },
+            },
+          }),
+          async () => {
+            try {
+              if (!this.imageCache.has(imageId)) {
+                await updateImageCache({
+                  imageCache: this.imageCache,
+                  imageElements: [imageElement],
+                  files: this.state.files,
+                });
+              }
+              if (
+                this.state.pendingImageElement?.id !== imageElement.id &&
+                this.state.draggingElement?.id !== imageElement.id
+              ) {
+                this.initializeImageDimensions(imageElement, true);
+              }
+              resolve(imageElement);
+            } catch (error) {
+              console.error(error);
+              reject(new Error("Couldn't create image"));
+            }
+          },
+        );
+      },
+    );
+  };
+
+  /**
+   * inserts image into elements array and rerenders
+   */
+  private insertImageElement = (
+    imageElement: ExcalidrawImageElement,
+    imageFile: File,
+  ) => {
+    this.scene.replaceAllElements([
+      ...this.scene.getElementsIncludingDeleted(),
+      imageElement,
+    ]);
+
+    this.initializeImage({
+      imageFile,
+      imageElement,
+    });
+
+    this.scene.informMutation();
+  };
+
+  private setImagePreviewCursor = async (imageFile: File) => {
+    setCursor(this.canvas, "wait");
+    const imageCompression = (await import("browser-image-compression"))
+      .default;
+    const imagePreview = await imageCompression(imageFile, {
+      maxWidthOrHeight: 100,
+      maxIteration: 1,
+    });
+    const previewDataURL = await getDataURL(imagePreview);
+    if (this.state.pendingImageElement) {
+      setCursor(this.canvas, `url(${previewDataURL}) 4 4, auto`);
+    }
+  };
+
+  private onImageAction = async (
+    { insertOnCanvasDirectly } = { insertOnCanvasDirectly: false },
+  ) => {
+    try {
+      const clientX = this.state.width / 2 + this.state.offsetLeft;
+      const clientY = this.state.height / 2 + this.state.offsetTop;
+
+      const { x, y } = viewportCoordsToSceneCoords(
+        { clientX, clientY },
+        this.state,
+      );
+
+      const imageFile = await fileOpen({
+        description: "Image",
+        extensions: [".jpg", ".jpeg", ".png", ".svg"],
+        mimeTypes: ["image/jpeg", "image/png", "image/svg+xml"],
+      });
+
+      this.setImagePreviewCursor(imageFile);
+
+      const imageElement = this.createImageElement({
+        sceneX: x,
+        sceneY: y,
+      });
+
+      if (insertOnCanvasDirectly) {
+        this.insertImageElement(imageElement, imageFile);
+        this.initializeImageDimensions(imageElement);
+        this.setState(
+          {
+            selectedElementIds: { [imageElement.id]: true },
+          },
+          () => {
+            this.actionManager.executeAction(actionFinalize);
+          },
+        );
+      } else {
+        this.setState(
+          {
+            pendingImageElement: imageElement,
+          },
+          () => {
+            this.insertImageElement(imageElement, imageFile);
+          },
+        );
+      }
+    } catch (error) {
+      if (error.name !== "AbortError") {
+        console.error(error);
+      }
+      this.setState(
+        {
+          pendingImageElement: null,
+          editingElement: null,
+          elementType: "selection",
+        },
+        () => {
+          this.actionManager.executeAction(actionFinalize);
+        },
+      );
+    }
+  };
+
+  private initializeImageDimensions = (
+    imageElement: ExcalidrawImageElement,
+    forceNaturalSize = false,
+  ) => {
+    const image =
+      isInitializedImageElement(imageElement) &&
+      this.imageCache.get(imageElement.imageId);
+
+    if (!image) {
+      if (
+        imageElement.width < DRAGGING_THRESHOLD / this.state.zoom.value &&
+        imageElement.height < DRAGGING_THRESHOLD / this.state.zoom.value
+      ) {
+        const defaultSize = 100;
+        mutateElement(imageElement, {
+          x: imageElement.x - defaultSize / 2,
+          y: imageElement.y - defaultSize / 2,
+          width: defaultSize,
+          height: defaultSize,
+        });
+      }
+
+      return;
+    }
+
+    if (
+      forceNaturalSize ||
+      // if user-created bounding box is below threshold, assume the
+      // intention was to click instead of drag, and use the image's
+      // intrinsic size
+      (imageElement.width < DRAGGING_THRESHOLD / this.state.zoom.value &&
+        imageElement.height < DRAGGING_THRESHOLD / this.state.zoom.value)
+    ) {
+      // some offset to account for UI
+      const offset = 160;
+
+      let width = Math.min(
+        image.naturalWidth,
+        this.state.width / this.state.zoom.value - offset,
+      );
+      let height = width * (image.naturalHeight / image.naturalWidth);
+
+      if (height > this.state.height / this.state.zoom.value - offset) {
+        height = this.state.height / this.state.zoom.value - offset;
+        width = height * (image.naturalWidth / image.naturalHeight);
+      }
+
+      // add current imageElement width/height to account for previous centering
+      // of the placholder image
+      const x = imageElement.x + imageElement.width / 2 - width / 2;
+      const y = imageElement.y + imageElement.height / 2 - height / 2;
+
+      mutateElement(imageElement, { x, y, width, height });
+    }
+  };
+
+  /** populates image cache and re-renders if needed */
+  private refreshImages = async (
+    imageElements: InitializedExcalidrawImageElement[] = getInitializedImageElements(
+      this.scene.getElements(),
+    ),
+    files: AppState["files"] = this.state.files,
+  ) => {
+    const uncachedImages = imageElements.filter(
+      (element) => !element.isDeleted && !this.imageCache.has(element.imageId),
+    );
+
+    if (uncachedImages.length) {
+      const { didUpdate } = await updateImageCache({
+        imageCache: this.imageCache,
+        imageElements: uncachedImages,
+        files,
+      });
+      if (didUpdate) {
+        this.scene.informMutation();
+      }
+    }
+  };
+
+  /** generally you should use `renderImages()` directly if you need to render
+   * new images. This is just a failsafe  */
+  private scheduleImageRefresh = throttle(() => {
+    this.refreshImages();
+  }, IMAGE_RENDER_TIMEOUT);
+
   private updateBindingEnabledOnPointerMove = (
     event: React.PointerEvent<HTMLCanvasElement>,
   ) => {
@@ -3889,31 +4285,59 @@ class App extends React.Component<AppProps, AppState> {
         }
       }
       const file = event.dataTransfer.files[0];
-      if (file?.type === "image/png" || file?.type === "image/svg+xml") {
-        if (fsSupported) {
+
+      if (isImageFile(file)) {
+        // first attempt to decode scene from the image if it's embedded
+        // ---------------------------------------------------------------------
+
+        if (file?.type === "image/png" || file?.type === "image/svg+xml") {
           try {
-            // This will only work as of Chrome 86,
-            // but can be safely ignored on older releases.
-            const item = event.dataTransfer.items[0];
-            (file as any).handle = await (item as any).getAsFileSystemHandle();
+            if (fsSupported) {
+              try {
+                // This will only work as of Chrome 86,
+                // but can be safely ignored on older releases.
+                const item = event.dataTransfer.items[0];
+                (file as any).handle = await (item as any).getAsFileSystemHandle();
+              } catch (error) {
+                console.warn(error.name, error.message);
+              }
+            }
+
+            const { elements, appState } = await loadFromBlob(
+              file,
+              this.state,
+              this.scene.getElementsIncludingDeleted(),
+            );
+            this.syncActionResult({
+              elements,
+              appState: {
+                ...(appState || this.state),
+                isLoading: false,
+              },
+              commitToHistory: true,
+            });
+            return;
           } catch (error) {
-            console.warn(error.name, error.message);
+            if (error.name !== "EncodingError") {
+              throw error;
+            }
           }
         }
 
-        const { elements, appState } = await loadFromBlob(
-          file,
+        // if no scene is embedded or we fail for whatever reason, fall back
+        // to importing as regular image
+        // ---------------------------------------------------------------------
+
+        const { x: sceneX, y: sceneY } = viewportCoordsToSceneCoords(
+          event,
           this.state,
-          this.scene.getElementsIncludingDeleted(),
         );
-        this.syncActionResult({
-          elements,
-          appState: {
-            ...(appState || this.state),
-            isLoading: false,
-          },
-          commitToHistory: true,
-        });
+
+        const imageElement = this.createImageElement({ sceneX, sceneY });
+        this.insertImageElement(imageElement, file);
+        this.initializeImageDimensions(imageElement);
+        this.setState({ selectedElementIds: { [imageElement.id]: true } });
+
         return;
       }
     } catch (error) {
@@ -4027,8 +4451,8 @@ class App extends React.Component<AppProps, AppState> {
         pointerCoords.y,
         distance(pointerDownState.origin.x, pointerCoords.x),
         distance(pointerDownState.origin.y, pointerCoords.y),
-        getResizeWithSidesSameLengthKey(event),
-        getResizeCenterPointKey(event),
+        shouldMaintainAspectRatio(event),
+        shouldResizeFromCenter(event),
       );
     } else {
       const [gridX, gridY] = getGridPoint(
@@ -4036,6 +4460,12 @@ class App extends React.Component<AppProps, AppState> {
         pointerCoords.y,
         this.state.gridSize,
       );
+
+      const image =
+        isInitializedImageElement(draggingElement) &&
+        this.imageCache.get(draggingElement.imageId);
+      const aspectRatio = image ? image.width / image.height : null;
+
       dragNewElement(
         draggingElement,
         this.state.elementType,
@@ -4045,9 +4475,13 @@ class App extends React.Component<AppProps, AppState> {
         gridY,
         distance(pointerDownState.originInGrid.x, gridX),
         distance(pointerDownState.originInGrid.y, gridY),
-        getResizeWithSidesSameLengthKey(event),
-        getResizeCenterPointKey(event),
+        draggingElement.type === "image"
+          ? !shouldMaintainAspectRatio(event)
+          : shouldMaintainAspectRatio(event),
+        shouldResizeFromCenter(event),
+        aspectRatio,
       );
+
       this.maybeSuggestBindingForAll([draggingElement]);
     }
   };
@@ -4080,9 +4514,11 @@ class App extends React.Component<AppProps, AppState> {
         transformHandleType,
         selectedElements,
         pointerDownState.resize.arrowDirection,
-        getRotateWithDiscreteAngleKey(event),
-        getResizeCenterPointKey(event),
-        getResizeWithSidesSameLengthKey(event),
+        shouldRotateWithDiscreteAngle(event),
+        shouldResizeFromCenter(event),
+        selectedElements.length === 1 && selectedElements[0].type === "image"
+          ? !shouldMaintainAspectRatio(event)
+          : shouldMaintainAspectRatio(event),
         resizeX,
         resizeY,
         pointerDownState.resize.center.x,